ability to skip some items in backup

+++ lib/backup/manager.rb
@@ -1,7 +1,5 @@
 module Backup
   class Manager
-    BACKUP_CONTENTS = %w{repositories/ db/ uploads/ backup_information.yml}
-
     def pack
       # saving additional informations
       s = {}
@@ -9,6 +7,7 @@ module Backup
       s[:backup_created_at]  = Time.now
       s[:gitlab_version]     = Gitlab::VERSION
       s[:tar_version]        = tar_version
+      s[:skipped]            = ENV["SKIP"]
       tar_file = "#{s[:backup_created_at].to_i}_gitlab_backup.tar"
 
       Dir.chdir(Gitlab.config.backup.path) do
@@ -17,12 +16,12 @@ module Backup
           file << s.to_yaml.gsub(/^---\n/,'')
         end
 
-        FileUtils.chmod(0700, %w{db uploads repositories})
+        FileUtils.chmod(0700, folders_to_backup)
 
         # create archive
         $progress.print "Creating backup archive: #{tar_file} ... "
         orig_umask = File.umask(0077)
-        if Kernel.system('tar', '-cf', tar_file, *BACKUP_CONTENTS)
+        if Kernel.system('tar', '-cf', tar_file, *backup_contents)
           $progress.puts "done".green
         else
           puts "creating archive #{tar_file} failed".red
@@ -46,6 +45,7 @@ module Backup
 
       connection = ::Fog::Storage.new(connection_settings)
       directory = connection.directories.get(remote_directory)
+
       if directory.files.create(key: tar_file, body: File.open(tar_file), public: false)
         $progress.puts "done".green
       else
@@ -56,7 +56,10 @@ module Backup
 
     def cleanup
       $progress.print "Deleting tmp directories ... "
-      BACKUP_CONTENTS.each do |dir|
+      
+      backup_contents.each do |dir|
+        next unless File.exist?(File.join(Gitlab.config.backup.path, dir))
+
         if FileUtils.rm_rf(File.join(Gitlab.config.backup.path, dir))
           $progress.puts "done".green
         else
@@ -73,6 +76,7 @@ module Backup
 
       if keep_time > 0
         removed = 0
+        
         Dir.chdir(Gitlab.config.backup.path) do
           file_list = Dir.glob('*_gitlab_backup.tar')
           file_list.map! { |f| $1.to_i if f =~ /(\d+)_gitlab_backup.tar/ }
@@ -84,6 +88,7 @@ module Backup
             end
           end
         end
+
         $progress.puts "done. (#{removed} removed)".green
       else
         $progress.puts "skipping".yellow
@@ -96,6 +101,7 @@ module Backup
       # check for existing backups in the backup dir
       file_list = Dir.glob("*_gitlab_backup.tar").each.map { |f| f.split(/_/).first.to_i }
       puts "no backups found" if file_list.count == 0
+
       if file_list.count > 1 && ENV["BACKUP"].nil?
         puts "Found more than one backup, please specify which one you want to restore:"
         puts "rake gitlab:backup:restore BACKUP=timestamp_of_backup"
@@ -110,6 +116,7 @@ module Backup
       end
 
       $progress.print "Unpacking backup ... "
+
       unless Kernel.system(*%W(tar -xf #{tar_file}))
         puts "unpacking backup failed".red
         exit 1
@@ -117,7 +124,6 @@ module Backup
         $progress.puts "done".green
       end
 
-      settings = YAML.load_file("backup_information.yml")
       ENV["VERSION"] = "#{settings[:db_version]}" if settings[:db_version].to_i > 0
 
       # restoring mismatching backups can lead to unexpected problems
@@ -135,6 +141,30 @@ module Backup
     def tar_version
       tar_version, _ = Gitlab::Popen.popen(%W(tar --version))
       tar_version.force_encoding('locale').split("\n").first
+    end
+
+    def skipped?(item)
+      settings[:skipped] && settings[:skipped].include?(item)
+    end
+
+    private
+
+    def backup_contents
+      folders_to_backup + ["backup_information.yml"]
+    end
+
+    def folders_to_backup
+      folders = %w{repositories db uploads}
+
+      if ENV["SKIP"]
+        return folders.reject{ |folder| ENV["SKIP"].include?(folder) }
+      end
+
+      folders
+    end
+
+    def settings
+      @settings ||= YAML.load_file("backup_information.yml")
     end
   end
 end
